osAgnostic: encoding is an invalid argument for subprocess.run. It was removed from all calls.
Changed git cloning from an SSH url to an http url to correct the issue of the script not being able to access the private repos.

subprocess.run cannot accept a Path object as a variable, so the names of the directories where changed to basic strings.

The script currently gets caught on line 119. It doesn't crash, it just runs indefinitely.